moving decorator out of api
